fix: update new where parser to handle errors gracefully
